requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
dnspython==2.6.1
email_validator==2.1.1
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.14.0
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googletrans==3.1.0a0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.23.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.4
jsonfield==3.1.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.5
kubernetes==29.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
nest-asyncio==1.5.6
nodejs==0.1.1
npm==0.1.1
numpy==1.25.0
oauthlib==3.2.2
onnxruntime==1.17.3
openai==0.28.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
optional-django==0.1.0
orjson==3.10.3
overrides==7.7.0
packaging==23.0
pandas==2.2.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.2.0
posthog==3.5.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.23.3
psutil==5.9.4
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.15.0
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.2
sympy==1.12
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tokenizers==0.19.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.6
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.6
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.18.2
